requirements.txt
numpy>=1.19.5
tensorflow>=2.5.0
scikit-learn>=0.24.2
pycryptodome>=3.10.1
matplotlib>=3.4.3 